Fix module outputs being unknown when changing organization structure (#14)
When the module is used as a resource, the evaluation of the `try` function causes all outputs that are based on `local.output_map` to become `(known after apply)` when the OU structure is changed. Removing the `try` function enables the module to output the expected structures with specific values as unknown.